Script is reworked to check invalid ImageType parameter
